requirements.txt
fpdf
python-dotenv
jwt
boto3
requests
flask
paramiko